Renames the project from todo to jot.
I decided to rename todo, mostly because it diverged from the
initial goal of being a todo-list manager.  I'm building it to
scratch an itch and I've realized that I never wanted a todo-list
manager--just a note-manager.

This commit replaces all instances of 'todo' with 'jot'.

The README.md has also been rewritten.